Generate debian package
- Build kernel normally
- Copy compiled files to output directory
- Create version and build debian package

For the debianization, mostly based on rpi repo:
https://github.com/RPi-Distro/firmware/blob/pios/bullseye/debian

Several details adapted and simplified for our use case.
Currently only build 64 bit kernel (kernel8). Compiling 32 bit kernel
will fail during package creation.

We are relying on the original rpikernelhack, which is needed to update
kernel and overlay files on /boot, because it's a fat32 partition and
dpkg can't update files there. For more info:
https://raspberrypi.stackexchange.com/a/94827

Signed-off-by: Daniel Finimundi <daniel@finimundi.com>