Build package with sbuild for cross-compiling
Since we want to make this package for arm archs, we need to cross
compile it. This can be easily done with the project
sbuild-debian-package.
Adding this as a submodule here to allow easier reproducibility on local
computers instead of just on CI environment.

FYI, official RPi image from RPi-Distro/firmware is built locally on a
Raspberry Pi to avoid cross compilation issues. See:
https://github.com/RPi-Distro/firmware/issues/7

Signed-off-by: Daniel Finimundi <daniel@finimundi.com>